Clean warnings.

+++ source/src/GraphicWindow.cc
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-GraphicWindow::GraphicWindow(size_t x, size_t y, std::string const& title) :
+GraphicWindow::GraphicWindow(uint_fast32_t x, uint_fast32_t y, std::string const& title) :
     window(sf::VideoMode(x, y), title)
 {
     window.clear();
